pyproject.toml
[project]
name = "face-rnd"
version = "0.1.0"
description = "Face Verification System using InsightFace and FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "insightface>=0.7.3",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "pydantic-settings>=2.10.1",
    "streamlit>=1.49.1",
    "plotly>=5.0.0",
    "pandas>=2.0.0",
    "aiohttp>=3.12.15",
    "onnxruntime-gpu>=1.22.0",
    "requests>=2.31.0",
    "torch>=2.0.0",
    "psutil>=7.0.0",
    "stringzilla==3.12.6",
    "qdrant-client>=1.15.1",
    "faiss-cpu>=1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

